requirements.txt
pytest==2.7.2
pyhamcrest==1.8.5
mock==1.3.0
